Fixed history data constructure to focus the target histroy
