+++ src/component/square.jsx
@@ -1,13 +1,8 @@
 import React from 'react';
 
-
 function Square({element, handleSquare}) {
-  function handleBoard() {
-    !element && handleSquare();
-  }
-
   return (
-    <button onClick={handleBoard} className="square">
+    <button onClick={handleSquare} className="square">
       {element}
     </button>
   )
